Tests for `createStyleBuilder().styled`
