+++ package.json
@@ -10,14 +10,15 @@
   },
   "devDependencies": {
     "@testing-library/react-hooks": "^8.0.0",
+    "@testing-library/react-native": "^9.1.0",
     "@types/react": "^18.0.12",
     "@types/react-native": "^0.67.8",
     "@typescript-eslint/eslint-plugin": "^5.26.0",
     "@typescript-eslint/parser": "^5.26.0",
+    "concurrently": "^7.2.2",
     "eslint": "^8.16.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "concurrently": "^7.2.2",
     "prettier": "^2.6.2",
     "react": "^18.1.0",
     "react-native": "^0.68.2",
